+++ app/templates/js/plugins.js
@@ -20,25 +20,3 @@
         }
     }
 }());
-
-// private Fast Selector wrapper,
-// returns jQuery object. Only use where
-// getElementById is not available.
-var S = function (selector, context) {
-	if (typeof selector === 'string') {
-		if (context) {
-			var cont;
-			if (context.jquery) {
-				cont = context[0];
-				if (!cont) return context;
-			} else {
-				cont = context;
-			}
-			return $(cont.querySelectorAll(selector));
-		}
-
-		return $(document.querySelectorAll(selector));
-	}
-
-	return $(selector, context);
-};